feat: add solution for modules
